requirements.txt
blinker==1.4
Jinja2==2.11.2
jsonify==0.5
MarkupSafe==1.1.1
marshmallow==3.6.1
PyJWT==1.7.1
PyYAML==5.3.1
six==1.15.0
Werkzeug==1.0.1

